+++ packages/picasso/src/SidebarItem/SidebarItem.tsx
@@ -189,6 +189,7 @@ export const getSelectedSubMenu = (sidebarItem: ReactElement<Props>) => {
   }
 
   const subMenuItems = React.Children.toArray(menu.props.children)
+
   return subMenuItems.find(
     (menuChild: ReactElement<Props>) => menuChild.props.selected
   )