+++ packages/shared/src/Picasso/config/test.ts
@@ -11,11 +11,13 @@ describe('responsive breakpoint utils', () => {
   describe('media query generation', () => {
     test('small', () => {
       const mediaQuery = screens('small')
+
       expect(mediaQuery).toEqual('@media (max-width: 576px)')
     })
 
     test('small medium', () => {
       const mediaQuery = screens('small', 'medium')
+
       expect(mediaQuery).toEqual(
         '@media (max-width: 576px), (min-width: 576px) and (max-width: 768px)'
       )
@@ -23,6 +25,7 @@ describe('responsive breakpoint utils', () => {
 
     test('small medium large', () => {
       const mediaQuery = screens('small', 'medium', 'large')
+
       expect(mediaQuery).toEqual(
         '@media (max-width: 576px), (min-width: 576px) and (max-width: 768px), (min-width: 768px) and (max-width: 992px)'
       )
@@ -32,56 +35,67 @@ describe('responsive breakpoint utils', () => {
   describe('screen size checks', () => {
     test('small breakpoint no screen size', () => {
       const isSmall = isScreenSize('small')
+
       expect(isSmall).toBeFalsy()
     })
 
     test('small breakpoint on a small screen', () => {
       const isSmall = isScreenSize('small', SCREEN_SIZES.small)
+
       expect(isSmall).toBeTruthy()
     })
 
     test('small breakpoint on a large screen', () => {
       const isSmall = isScreenSize('small', SCREEN_SIZES.large)
+
       expect(isSmall).toBeFalsy()
     })
 
     test('medium breakpoint no screen size', () => {
       const isMedium = isScreenSize('medium')
+
       expect(isMedium).toBeFalsy()
     })
 
     test('medium breakpoint on a small screen', () => {
       const isMedium = isScreenSize('medium', SCREEN_SIZES.small)
+
       expect(isMedium).toBeFalsy()
     })
 
     test('medium breakpoint on a medium screen', () => {
       const isMedium = isScreenSize('medium', SCREEN_SIZES.medium)
+
       expect(isMedium).toBeTruthy()
     })
 
     test('medium breakpoint on a large screen', () => {
       const isMedium = isScreenSize('medium', SCREEN_SIZES.large)
+
       expect(isMedium).toBeFalsy()
     })
 
     test('large breakpoint on a medium screen', () => {
       const isLarge = isScreenSize('large', SCREEN_SIZES.medium)
+
       expect(isLarge).toBeFalsy()
     })
 
     test('large breakpoint on a large screen', () => {
       const isLarge = isScreenSize('large', SCREEN_SIZES.large)
+
       expect(isLarge).toBeTruthy()
     })
 
     test('extra large breakpoint on a  large screen', () => {
       const isExtraLarge = isScreenSize('extra-large', SCREEN_SIZES.large)
+
       expect(isExtraLarge).toBeFalsy()
     })
 
     test('extra large breakpoint on a extra large screen', () => {
       const isExtraLarge = isScreenSize('extra-large', SCREEN_SIZES.extraLarge)
+
       expect(isExtraLarge).toBeTruthy()
     })
   })
@@ -95,16 +109,19 @@ describe('non-responsive breakpoint utils', () => {
   describe('media query generation', () => {
     test('small', () => {
       const mediaQuery = screens('small')
+
       expect(mediaQuery).toEqual('')
     })
 
     test('small medium', () => {
       const mediaQuery = screens('small', 'medium')
+
       expect(mediaQuery).toEqual('')
     })
 
     test('small medium large', () => {
       const mediaQuery = screens('small', 'medium', 'large')
+
       expect(mediaQuery).toEqual(
         '@media (min-width: 768px) and (max-width: 992px)'
       )
@@ -114,56 +131,67 @@ describe('non-responsive breakpoint utils', () => {
   describe('screen size checks', () => {
     test('small breakpoint no screen size', () => {
       const isSmall = isScreenSize('small')
+
       expect(isSmall).toBeFalsy()
     })
 
     test('small breakpoint on a small screen', () => {
       const isSmall = isScreenSize('small', SCREEN_SIZES.small)
+
       expect(isSmall).toBeTruthy()
     })
 
     test('small breakpoint on a large screen', () => {
       const isSmall = isScreenSize('small', SCREEN_SIZES.large)
+
       expect(isSmall).toBeFalsy()
     })
 
     test('medium breakpoint no screen size', () => {
       const isMedium = isScreenSize('medium')
+
       expect(isMedium).toBeFalsy()
     })
 
     test('medium breakpoint on a small screen', () => {
       const isMedium = isScreenSize('medium', SCREEN_SIZES.small)
+
       expect(isMedium).toBeFalsy()
     })
 
     test('medium breakpoint on a medium screen', () => {
       const isMedium = isScreenSize('medium', SCREEN_SIZES.medium)
+
       expect(isMedium).toBeFalsy()
     })
 
     test('medium breakpoint on a large screen', () => {
       const isMedium = isScreenSize('medium', SCREEN_SIZES.large)
+
       expect(isMedium).toBeFalsy()
     })
 
     test('large breakpoint on a medium screen', () => {
       const isLarge = isScreenSize('large', SCREEN_SIZES.medium)
+
       expect(isLarge).toBeFalsy()
     })
 
     test('large breakpoint on a large screen', () => {
       const isLarge = isScreenSize('large', SCREEN_SIZES.large)
+
       expect(isLarge).toBeTruthy()
     })
 
     test('extra large breakpoint on a  large screen', () => {
       const isExtraLarge = isScreenSize('extra-large', SCREEN_SIZES.large)
+
       expect(isExtraLarge).toBeFalsy()
     })
 
     test('extra large breakpoint on a extra large screen', () => {
       const isExtraLarge = isScreenSize('extra-large', SCREEN_SIZES.extraLarge)
+
       expect(isExtraLarge).toBeTruthy()
     })
   })