+++ packages/shared/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.16.5](https://github.com/toptal/picasso/compare/@toptal/picasso-shared@1.16.4...@toptal/picasso-shared@1.16.5) (2020-10-27)
+
+**Note:** Version bump only for package @toptal/picasso-shared
+
+
+
+
+
 ## [1.16.4](https://github.com/toptal/picasso/compare/@toptal/picasso-shared@1.16.3...@toptal/picasso-shared@1.16.4) (2020-09-09)
 
 **Note:** Version bump only for package @toptal/picasso-shared